// front/src/models/bundesland.enum.ts
export enum Bundesland {
    Vorarlberg = "Vorarlberg",
    Tirol = "Tirol",
    Salzburg = "Salzburg",
    Kärnten = "Kärnten",
    Oberösterreich = "Oberösterreich",
    Niederösterreich = "Niederösterreich",
    Steiermark = "Steiermark",
    Wien = "Wien",
    Burgenland = "Burgenland"  
}  